Added Curl and Rapidjson
Didn't implement it into the code yet but added the files and made a separate file for Shipping cost for a future project of implementing an API that can detect location being entered. The code would then take the distances between both location and do a rate based off of KM, or maybe a flat $8.44 like UPS international shipping.

- Also adding .ipch files to .gitignore due to file size